Backend/requirements.txt
Django==5.2.6
djangorestframework==3.16.1
psycopg2-binary==2.9.10
python-decouple==3.8
dj-database-url==2.1.0
gunicorn==21.2.0
whitenoise==6.6.0
django-cors-headers==4.3.1
django-filter==24.2
django-allauth==0.57.0
requests==2.31.0
